Include Strict Transport Security on error pages
In our "lets-nginx" deployment we added basic auth, and so the SSL Labs client (or any unauthenticated client) was receiving 401s. By default, [nginx ignores "add_header" directives on error pages](http://serverfault.com/a/418717), and so we downgraded to a mere A grade because we weren't sending back the Strict-Transport-Security header on errors.

This change to the nginx configuration ensures an A+ grade even when serving error pages (and also closes an extremely narrow window of downgrade attack around the same). I'm not familiar with the other two headers, so I'm not sure if the same change should be applied to them as well – if so, just let me know and I'll be happy to update the PR.