Added the possibility to configure multiple destinations, according to the domain name. You can now run the container with `-e DOMAIN="domain1.com;domain2.com"` The UPSTREAM env value need to specify the different backend services, 1 per domain.
I also updated the FROM directory in the dockerfile, in order to pull the latest version of alpine.
Bash is also installed to halp handle the arrays needed in the script.

Few changes have also been made in the letsencrypt command, which was outdated (the --authenticator option is no longer recognized, as well as --renew by default).
I tested this on my machine, it's building and running fine.
I hope this help, please let me know if you have any issues :)